fix: move getAllDapps where it belongs
