+++ src/pages/index.tsx
@@ -4,9 +4,9 @@ import Categories from "../components/Categories/Categories"
 import FilterMenu from "../components/FilterMenu/FilterMenu"
 import Layout from "../components/Layout"
 import Select from "../components/Select/Select"
+import { getAllDapps } from "../data/getAllDapps"
 import { filterDappcardsByRating, getRatings } from "../helpers/rating"
 import sortByAttribute from "../helpers/sort"
-import { getAllDapps } from "../hooks/getAllDapps"
 import { useCategoryStore } from "../hooks/useCategoryStore"
 import { useRouter } from "next/router"
 import { useEffect, useState } from "react"
